Add decorator and composition
