+++ capitalize_decorator.rb
@@ -1,0 +1,6 @@
+require_relative 'decorator'
+class CapitalizeDecorator < Decorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
