+++ decorator.rb
@@ -1,0 +1,11 @@
+require_relative 'nameable'
+class Decorator < Nameable
+  def initialize(nameable)
+    @nameable = nameable
+    super()
+  end
+
+  def correct_name
+    @nameable.correct_name
+  end
+end
